made 1 file for all graph views
