+++ myapp/urls.py
@@ -6,11 +6,8 @@ from myapp.Views.currencyNewsView import CurrencyNewsListView, FetchCurrencyNews
 from myapp.Views.exchangeRateLatestViews import CurrencyRateView
 from myapp.Views.historicalExchangeRatesView import FetchHistoricalCurrencyExchangeRates
 from myapp.Views.volatilityView import VolatilityAnalysisView
-from myapp.Views.graphView_lastweek import GraphView_lastweek
-from myapp.Views.graphView_lastmonth import GraphView_lastmonth
-from myapp.Views.graphView_last6months import GraphView_last6months
-from myapp.Views.graphView_lastyear import GraphView_lastyear
-from myapp.Views.graphView_last5years import GraphView_last5years
+from myapp.Views.graphView import (GraphView_lastweek, GraphView_lastmonth,
+                                   GraphView_last6months, GraphView_lastyear, GraphView_last5years)
 from myapp.Views.registerExchangeRateAlertView import RegisterAlertView
 from myapp.Views.anomalyDetectionView import anomaly_detection
